# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateExam:
    properties:
      allowedTime:
        type: integer
      curriculum:
        type: integer
      departement:
        type: string
      subject:
        type: string
      year:
        type: integer
    required:
    - allowedTime
    - curriculum
    - departement
    - subject
    - year
    type: object
  dto.CreateQuestion:
    properties:
      answer:
        type: string
      choose:
        additionalProperties:
          type: string
        type: object
      description:
        type: string
      exam_id:
        type: integer
      image_url:
        type: string
      question:
        type: string
      question_number:
        type: integer
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.ExamAnswers:
    properties:
      answer:
        type: string
      question_id:
        type: integer
    required:
    - answer
    - question_id
    type: object
  dto.ExamResponse:
    properties:
      allowedTime:
        type: integer
      curriculum:
        type: integer
      departement:
        type: string
      id:
        type: integer
      question:
        items:
          $ref: '#/definitions/dto.QuestionResponse'
        type: array
      subject:
        type: string
      year:
        type: integer
    type: object
  dto.MockExam:
    properties:
      answer:
        items:
          $ref: '#/definitions/dto.ExamAnswers'
        type: array
    type: object
  dto.QuestionResponse:
    properties:
      answer:
        type: string
      choose:
        additionalProperties:
          type: string
        type: object
      description:
        type: string
      exam:
        $ref: '#/definitions/dto.ExamResponse'
      id:
        type: integer
      imageUrl:
        type: string
      question:
        type: string
      question_Number:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Exam API
  title: Exam API
  version: "1.0"
paths:
  /exam:
    get:
      description: Get list of all exams
      produces:
      - application/json
      responses:
        "200":
          description: All exams
          schema:
            items:
              $ref: '#/definitions/dto.ExamResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all exams
      tags:
      - Exam
    post:
      consumes:
      - application/json
      description: Create a new exam with subject, department, year, curriculum and
        allowed time
      parameters:
      - description: Create Exam
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExam'
      produces:
      - application/json
      responses:
        "200":
          description: Exam created successfully
          schema:
            $ref: '#/definitions/dto.ExamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new Exam
      tags:
      - Exam
  /exam/:
    get:
      description: Get exam details with paginated questions based on query parameters
      parameters:
      - description: Subject
        in: query
        name: subject
        type: string
      - description: Department
        in: query
        name: departement
        type: string
      - description: Year
        in: query
        name: year
        type: integer
      - description: Curriculum
        in: query
        name: curriculum
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 2
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: question_number
        description: Sort field
        in: query
        name: sort
        type: string
      - default: asc
        description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Exam found
          schema:
            $ref: '#/definitions/dto.ExamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get exam with questions
      tags:
      - Exam
  /exam/{id}:
    delete:
      description: Delete exam by ID
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Exam deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete exam
      tags:
      - Exam
    put:
      consumes:
      - application/json
      description: Update exam details by ID
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Exam
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExam'
      produces:
      - application/json
      responses:
        "200":
          description: Exam updated successfully
          schema:
            $ref: '#/definitions/dto.ExamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update exam
      tags:
      - Exam
  /exam/score:
    post:
      consumes:
      - application/json
      description: Calculate total score based on multiple question answers
      parameters:
      - description: Mock exam answers
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MockExam'
      produces:
      - application/json
      responses:
        "200":
          description: Total score
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Calculate total score for multiple questions
      tags:
      - Exam
  /exam/score/{id}/{answer}:
    get:
      description: Calculate if the answer for a specific question is correct
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      - description: Answer
        in: path
        name: answer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Score calculated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Calculate score for one question
      tags:
      - Exam
  /question:
    get:
      description: All Question as a lis
      produces:
      - application/json
      responses:
        "200":
          description: ALL Question Founded
          schema:
            items:
              $ref: '#/definitions/dto.QuestionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: GetAll Question
      tags:
      - Questions
  /question/{id}:
    delete:
      description: Delete Question by ID
      parameters:
      - description: Question id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Question Deleted Successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete Question
      tags:
      - Questions
    get:
      description: use question id to get only one question
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Question found
          schema:
            $ref: '#/definitions/dto.QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get Question by ID
      tags:
      - Questions
    put:
      consumes:
      - application/json
      description: Update Question by ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Question
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateQuestion'
      produces:
      - application/json
      responses:
        "200":
          description: Question updated successfully
          schema:
            $ref: '#/definitions/dto.QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update Question
      tags:
      - Questions
  /question/create:
    post:
      consumes:
      - application/json
      description: add a new exam question
      parameters:
      - description: 'Created Question '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateQuestion'
      produces:
      - application/json
      responses:
        "200":
          description: 'Question created successfully '
          schema:
            $ref: '#/definitions/dto.QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new Question
      tags:
      - Questions
swagger: "2.0"